Add typescript, first part of cleanup code, add styled components.
