+++ package.json
@@ -9,6 +9,15 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "@types/core-js": "^2.5.4",
+    "@types/lodash": "^4.14.168",
+    "@types/node": "^14.14.37",
+    "@types/react": "^17.0.3",
+    "@types/react-dom": "^17.0.3",
+    "@types/react-router-dom": "^5.1.7",
+    "@types/styled-components": "^5.1.9",
+    "@types/webpack-env": "^1.16.0",
+    "core-js": "^3.10.0",
     "dotenv": "^8.2.0",
     "emailjs-com": "^2.6.4",
     "env-cmd": "^10.1.0",
@@ -31,8 +40,12 @@
     "react-scripts": "4.0.3",
     "reactjs-popup": "^2.0.4",
     "require-context": "^1.1.0",
+    "styled-components": "^5.2.3",
+    "tslint": "^6.1.3",
+    "typescript": "^4.2.3",
     "valid-url": "^1.0.9",
-    "web-vitals": "^1.0.1"
+    "web-vitals": "^1.0.1",
+    "webpack-env": "^0.8.0"
   },
   "scripts": {
     "dev": "react-scripts start",
