+++ src/entities/ProjectEntity.ts
@@ -1,0 +1,6 @@
+export interface ISingleProject {
+    id: string;
+    language: string;
+    owner: string;
+    name: string;
+}
